% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactive_prospects.R
\name{inactive_prospects}
\alias{inactive_prospects}
\title{Inactive Prospects Report}
\usage{
inactive_prospects()
}
\description{
Output list of assigned prospects who have not been contacted in the last year
}
\examples{
## run reports as of today
inactive_prospects()
}
